maintaining a database of "nominal" values doesn't make sense in the era of multiple GENIE CMCs
Since GENIE 3, GENIE supports multiple "comprehensive MC configurations," which have different combinations of active models and tuned parameters.  That means that unless we want to store the nominal values for each of these (we don't), the "nominal" no longer makes sense.  There are also a few parameters for which there's no fixed "nominal" value.  So here we simply remove this functionality.  I don't believe anybody is using it anyway.